open_entity() return value issue fixed. if the object doesn't exist, `open_entity()` returns `-ENOENT` (`-2`) if the object exists, `open_entity()` successes in opening the object(reading object attributes back from ioservices), it returns 0

+++ c0cat.c
@@ -27,8 +27,8 @@
 /* main */
 int main(int argc, char **argv)
 {
-	int idh;	/* object id high 	*/
-	int idl;	/* object is low	*/
+	int64_t idh;/* object id high 	*/
+	int64_t idl;/* object is low	*/
 	int bsz;	/* block size 		*/
 	int cnt;	/* count			*/
 
@@ -45,8 +45,8 @@ int main(int argc, char **argv)
 	sprintf(c0rcfile,".%src",basename(argv[0]));
 
 	/* set input */
-	idh = atoi(argv[1]);
-	idl = atoi(argv[2]);
+	idh = atoll(argv[1]);
+	idl = atoll(argv[2]);
 	bsz = atoi(argv[3]);
 	cnt = atoi(argv[4]);
